Migrated to using buckminster plugin:
>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>

Latest changes to the p2 site we use to populate the tp/ (target
platform) directory when you materialize Dawn/GDA/IDA etc using
Buckminster. 

Eclipse components:
•	No changes

Non-Eclipse components:
(1) Newer versions of components that were already available, with
bundle name unchanged (previous versions are still available)
•	No changes

(2) Newly added components
•	uk.ac.diamond.org.mihalis.opal  – version 0.9.5.2 is available. New
widgets for the SWT API.

The list of components that can be materialized is in the Infrastructure
Guide and has been updated with the above.

Note well: For many components, all versions are published to the same
p2 site (from where the target platform is populated). Therefore, where
your plugins declare a dependency on a tp/ component, you *must* specify
an exact version (or a version range). If you don’t, then a future
materialize might unexpectedly get a different, later, version of the
component. Note that this is best practice for declaring dependencies. 

Matthew