HDF5/Nexus Workspace and File model
A caching model that is used by Eclipse to accelerate sorting
and filtering of large groups of Nexus (and HDF5) files.

Includes tests that use Mockito, which temporarily means that
this plug-in depends on Mockito/Hamcrest/Objenesis. At some
future point the pre-existing and new tests can be split into
a fragment.