SCI-9675 Update Apache Batik libraries to 1.9.1
This allows org.dawb.common.ui to compile with Java 11 compliance. Add
workaround for missing services

These libraries are problematic in that the Eclipse release p2 provides
multiple versions hence the unusual practice of including the specific
version in the feature.xml. This means that the products include two
versions of some of the libraries.

It would be better if we could use the latest version from the Eclipse
p2 (which is 1.11 for 2020-03) however there is some inconsistency in
which modules are published.

However, looking at the modules in the 2021-03 Eclipse Orbit, it looks
promising that the available 1.13 versions of Batik there will be
compatible with our requirements so we should revisit after the target
platform is advanced.

Change-Id: I21dc7ffdb49c876754bbf8c99b5a5a7cd0eb33b9